+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-08-20 17:38:16
-// Commit: chore(test): update dependencies and cleanup unused imports in test
+// Updated at: 2026-08-20 17:58:47
+// Commit: perf(test): optimize memory consumption in test list viewport
 
 class testComponentState {
-  final String commitRef = '3773';
-  final String lastMessage = 'chore(test): update dependencies and cleanup unused imports in test';
-  final DateTime timestamp = DateTime.parse('2026-08-20T17:38:16+05:00');
+  final String commitRef = '3774';
+  final String lastMessage = 'perf(test): optimize memory consumption in test list viewport';
+  final DateTime timestamp = DateTime.parse('2026-08-20T17:58:47+05:00');
 }